+++ layout/src/main/java/com/vlkan/log4j2/logstash/layout/LogstashLayout.java
@@ -294,7 +294,7 @@ public class LogstashLayout implements Layout<String> {
         private boolean stackTraceEnabled = false;
 
         @PluginBuilderAttribute
-        private boolean emptyPropertyExclusionEnabled = true;
+        private boolean emptyPropertyExclusionEnabled = false;
 
         @PluginBuilderAttribute
         private String dateTimeFormatPattern = "yyyy-MM-dd'T'HH:mm:ss.SSSZZZ";